[HOTSPOT-181] 구성원별 데이터 한도 및 즉시 차단 여부 업데이트 API 수정

+++ src/main/java/hotspot/user/family/infrastructure/entity/FamilyEntity.java
@@ -44,7 +44,7 @@ public class FamilyEntity extends BaseEntity {
 
     @Column(nullable = false)
     @Builder.Default
-    private int familyDataAmount = 0;
+    private long familyDataAmount = 0;
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
